Created new cli file for run methods
